+++ src/app/components/maintenance/application-type.component.ts
@@ -1,6 +1,6 @@
 import {Component, EventEmitter, Input, Output, ViewChild} from '@angular/core';
 import {ModalService, ShareDataService} from '@win-angular/services';
-import {ApplicationType} from '../../model';
+import {ApplicationType, Permissions} from '../../model';
 import {ApplicationTypeService} from '../../services';
 
 @Component({
@@ -11,6 +11,7 @@ import {ApplicationTypeService} from '../../services';
 export class ApplicationTypeComponent {
 
   @Input() modalId: string;
+  @Input() permissions: Permissions;
   @Output() createEvent: EventEmitter<ApplicationType> = new EventEmitter<ApplicationType>();
   @Output() deleteEvent: EventEmitter<ApplicationType> = new EventEmitter<ApplicationType>();
   @Output() updateEvent: EventEmitter<ApplicationType> = new EventEmitter<ApplicationType>();
@@ -38,6 +39,10 @@ export class ApplicationTypeComponent {
     this.newApplicationTypeForm.resetForm();
   }
 
+  public hasAdminPermissions(): boolean {
+    return this.permissions.permissions.indexOf('APM_Admin') >= 0;
+  }
+
   public saveApplicationType(): void {
     if (this.passedApplicationType) {
       this.updateApplicationType();
@@ -48,6 +53,11 @@ export class ApplicationTypeComponent {
 
   public createApplicationType(): void {
     const created: ApplicationType = Object.assign({}, this.model);
+
+    if (!this.hasAdminPermissions()) {
+      this.shareDataService.showStatus([{severity: 'error', summary: 'You are not authorized to create an application type'}]);
+      return;
+    }
 
     this.applicationTypeService.create(created)
       .subscribe(res => {
@@ -63,6 +73,11 @@ export class ApplicationTypeComponent {
   public updateApplicationType(): void {
     const updated: ApplicationType = Object.assign({}, this.model);
 
+    if (!this.hasAdminPermissions()) {
+      this.shareDataService.showStatus([{severity: 'error', summary: 'You are not authorized to update an application type'}]);
+      return;
+    }
+
     this.applicationTypeService.update(updated)
       .subscribe(res => {
           this.closeModal();
@@ -76,6 +91,11 @@ export class ApplicationTypeComponent {
 
   public deleteApplicationType(): void {
     const deleted: ApplicationType = Object.assign({}, this.model);
+
+    if (!this.hasAdminPermissions()) {
+      this.shareDataService.showStatus([{severity: 'error', summary: 'You are not authorized to delete an application type'}]);
+      return;
+    }
 
     this.applicationTypeService.delete(deleted)
       .subscribe(res => {
